refactor: Voucher 추상클래스화
