+++ src/main/java/com/devcourse/springbootbasic/engine/voucher/domain/PercentDiscountVoucher.java
@@ -4,32 +4,15 @@ import com.devcourse.springbootbasic.engine.model.VoucherType;
 
 import java.util.UUID;
 
-public class PercentDiscountVoucher implements Voucher {
+public class PercentDiscountVoucher extends Voucher {
 
-    private final UUID voucherId;
-    private final VoucherType voucherType;
-    private final double discountPercent;
-
-    public PercentDiscountVoucher(UUID voucherId, VoucherType voucherType, double percent) {
-        this.voucherId = voucherId;
-        this.voucherType = voucherType;
-        this.discountPercent = percent;
+    PercentDiscountVoucher(UUID uuid, VoucherType voucherType, double percent) {
+        super(uuid, voucherType, percent);
     }
 
     @Override
     public double discountedPrice(long originalPrice) {
-        return originalPrice - (1 - discountPercent / (double) 100);
+        return originalPrice - (1 - discountValue / (double) 100);
     }
 
-    @Override
-    public String toString() {
-        return voucherType.name() + " -> id:" + voucherId +
-                ", discount" + voucherType.getTypeString() +
-                ":" + discountPercent;
-    }
-
-    @Override
-    public UUID getVoucherId() {
-        return voucherId;
-    }
 }
