+++ src/main/java/com/devcourse/springbootbasic/engine/voucher/domain/Voucher.java
@@ -1,11 +1,30 @@
 package com.devcourse.springbootbasic.engine.voucher.domain;
 
+import com.devcourse.springbootbasic.engine.model.VoucherType;
+
 import java.util.UUID;
 
-public interface Voucher {
-    double discountedPrice(long originalPrice);
+public abstract class Voucher {
+    UUID voucherId;
+    VoucherType voucherType;
+    double discountValue;
 
-    UUID getVoucherId();
+    Voucher(UUID voucherId, VoucherType voucherType, double discountAmount) {
+        this.voucherId = voucherId;
+        this.voucherType = voucherType;
+        this.discountValue = discountAmount;
+    }
 
-    String toString();
+    abstract double discountedPrice(long originalPrice);
+
+    UUID getVoucherId() {
+        return voucherId;
+    }
+
+    public String toString() {
+        return voucherType.name() + " -> " +
+                "id:" + voucherId +
+                ", discount" + voucherType.getTypeString() +
+                ":" + discountValue;
+    }
 }
